Move internal/otelutil to daemon/internal/otelutil
