+++ daemon/command/daemon.go
@@ -28,6 +28,7 @@ import (
 	buildkit "github.com/docker/docker/daemon/internal/builder-next"
 	"github.com/docker/docker/daemon/internal/builder-next/exporter"
 	"github.com/docker/docker/daemon/internal/libcontainerd/supervisor"
+	"github.com/docker/docker/daemon/internal/otelutil"
 	"github.com/docker/docker/daemon/listeners"
 	dopts "github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/daemon/pkg/plugin"
@@ -48,7 +49,6 @@ import (
 	systemrouter "github.com/docker/docker/daemon/server/router/system"
 	"github.com/docker/docker/daemon/server/router/volume"
 	"github.com/docker/docker/dockerversion"
-	"github.com/docker/docker/internal/otelutil"
 	"github.com/docker/docker/pkg/authorization"
 	"github.com/docker/docker/pkg/homedir"
 	"github.com/docker/docker/pkg/pidfile"
