+++ daemon/create.go
@@ -20,8 +20,8 @@ import (
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/internal/metrics"
 	"github.com/docker/docker/daemon/internal/multierror"
+	"github.com/docker/docker/daemon/internal/otelutil"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/otelutil"
 	"github.com/docker/docker/runconfig"
 	"github.com/moby/moby/api/types/backend"
 	containertypes "github.com/moby/moby/api/types/container"
