+++ daemon/daemon_unix.go
@@ -27,6 +27,7 @@ import (
 	"github.com/docker/docker/daemon/container"
 	"github.com/docker/docker/daemon/initlayer"
 	"github.com/docker/docker/daemon/internal/libcontainerd/remote"
+	"github.com/docker/docker/daemon/internal/otelutil"
 	"github.com/docker/docker/daemon/internal/usergroup"
 	"github.com/docker/docker/daemon/libnetwork"
 	nwconfig "github.com/docker/docker/daemon/libnetwork/config"
@@ -38,7 +39,6 @@ import (
 	"github.com/docker/docker/daemon/pkg/opts"
 	volumemounts "github.com/docker/docker/daemon/volume/mounts"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/otelutil"
 	"github.com/docker/docker/pkg/sysinfo"
 	"github.com/docker/docker/runconfig"
 	"github.com/moby/moby/api/types/blkiodev"
