+++ daemon/internal/otelutil/status.go
