+++ daemon/libnetwork/sandbox_externalkey_unix.go
@@ -13,8 +13,8 @@ import (
 	"path/filepath"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/otelutil"
 	"github.com/docker/docker/daemon/libnetwork/types"
-	"github.com/docker/docker/internal/otelutil"
 	"github.com/docker/docker/pkg/stringid"
 	"github.com/moby/sys/reexec"
 	"github.com/opencontainers/runtime-spec/specs-go"
