+++ daemon/server/server.go
@@ -5,12 +5,12 @@ import (
 	"net/http"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/otelutil"
 	"github.com/docker/docker/daemon/server/httpstatus"
 	"github.com/docker/docker/daemon/server/httputils"
 	"github.com/docker/docker/daemon/server/middleware"
 	"github.com/docker/docker/daemon/server/router"
 	"github.com/docker/docker/dockerversion"
-	"github.com/docker/docker/internal/otelutil"
 	"github.com/gorilla/mux"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/versions"
