+++ daemon/start.go
@@ -11,8 +11,8 @@ import (
 	mobyc8dstore "github.com/docker/docker/daemon/containerd"
 	"github.com/docker/docker/daemon/internal/libcontainerd"
 	"github.com/docker/docker/daemon/internal/metrics"
+	"github.com/docker/docker/daemon/internal/otelutil"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/otelutil"
 	"github.com/moby/moby/api/types/backend"
 	"github.com/moby/moby/api/types/events"
 	"github.com/pkg/errors"
